Switch to using virt-install plus more flexibility.
Switching to virt-install makes it easier to port to OSes other than Fedora, like RHEL.

In case you use another network interface on KVM/libvirt than the default one,
you can not specify that in the config file.

Also there is now a system index, so that you can build multiple COS systems on the
same hypervisor.

Finally increased the time to wait for the appliances to boot a bit to tolerate slower systems.